src/requirements.txt
Django==1.3.1
psycopg2==2.4
south
pil
django_classy_tags
django_cms
django_extensions
django_mptt
django_sekizai
django_userena
django_guardian
django_tagging
easy_thumbnails
html5lib
icalendar
django-filer >= 0.8.0
cmsplugin-filer
django-tagging
debug_toolbar